+++ tests/WordFormatter/LowerCaseConverterTest.php
@@ -9,18 +9,18 @@ use Stadly\PasswordPolice\WordFormatter;
 use Traversable;
 
 /**
- * @coversDefaultClass \Stadly\PasswordPolice\WordFormatter\LowerCase
+ * @coversDefaultClass \Stadly\PasswordPolice\WordFormatter\LowerCaseConverter
  * @covers ::<protected>
  * @covers ::<private>
  */
-final class LowerCaseTest extends TestCase
+final class LowerCaseConverterTest extends TestCase
 {
     /**
      * @covers ::apply
      */
     public function testCanFormatWord(): void
     {
-        $formatter = new LowerCase();
+        $formatter = new LowerCaseConverter();
 
         self::assertSame(['foobar'], iterator_to_array($formatter->apply(['fOoBaR']), false));
     }
@@ -30,7 +30,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testCanFormatWords(): void
     {
-        $formatter = new LowerCase();
+        $formatter = new LowerCaseConverter();
 
         self::assertSame(['foo', 'bar'], iterator_to_array($formatter->apply(['fOo', 'BaR']), false));
     }
@@ -40,7 +40,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testCanFormatUtf8Characters(): void
     {
-        $formatter = new LowerCase();
+        $formatter = new LowerCaseConverter();
 
         self::assertSame(['ááæøôëñ'], iterator_to_array($formatter->apply(['áÁæØôËñ']), false));
     }
@@ -50,7 +50,7 @@ final class LowerCaseTest extends TestCase
      */
     public function testCanApplyFormatterChain(): void
     {
-        $formatter = new LowerCase();
+        $formatter = new LowerCaseConverter();
 
         $next = $this->createMock(WordFormatter::class);
         $next->method('apply')->willReturnCallback(